+++ primitives/shared/src/lib.rs
@@ -5,4 +5,4 @@ pub mod piece;
 
 /// Merkle tree node size in bytes.
 /// TODO: Where should this be moved to?
-pub const NODE_SIZE: u64 = 32;
+pub const NODE_SIZE: usize = 32;
